Fix wrapped unary and binary operators
Using the "safe eval" trick worked in some cases, but not if any of the
operands was a string, for example. (I had code that said "if (!op)",
where op was "*", and so that led to evaluating "return ! *" which is a
syntax error.

The lists of operators come from the ones listed in the Babel AST
documentation as possible values for the "operator" property of
UnaryExpression and BinaryExpression:
https://babeljs.io/docs/en/babel-types#unaryexpression
https://babeljs.io/docs/en/babel-types#binaryexpression

I don't really like these switch statements where we have to make sure we
include every operator. The alternative is to add a dependency on a
package like https://www.npmjs.com/package/tosource and then also make it
available at the global scope, which doesn't seem great either. But then
we'd be able to do:

    Function(`return ${toSource(left)} ${op} ${toSource(right)}`)()

I think the switch statement is probably the less worse solution.